feat: add command exists checking for all tools installation

+++ linux/installation/developer-packages/ubuntu/tools/bandwhich.sh
@@ -1,7 +1,13 @@
 #! /bin/bash
 
-wget https://github.com/imsnif/bandwhich/releases/download/v0.23.1/bandwhich-v0.23.1-x86_64-unknown-linux-gnu.tar.gz
-tar -xzf bandwhich-v0.23.1-x86_64-unknown-linux-gnu.tar.gz
-sudo cp bandwhich /usr/local/bin/bandwhich
-echo "==== CLEAN UP ===="
-rm -f bandwhich-v0.23.1-x86_64-unknown-linux-gnu.tar.gz bandwhich
+if ! command -v bandwhich 2>&1 >/dev/null
+then
+    echo "[INSTALLING ⬇️ ] bandwhich"
+    wget https://github.com/imsnif/bandwhich/releases/download/v0.23.1/bandwhich-v0.23.1-x86_64-unknown-linux-gnu.tar.gz
+    tar -xzf bandwhich-v0.23.1-x86_64-unknown-linux-gnu.tar.gz
+    sudo cp bandwhich /usr/local/bin/bandwhich
+    echo "==== CLEAN UP ===="
+    rm -f bandwhich-v0.23.1-x86_64-unknown-linux-gnu.tar.gz bandwhich
+else
+    echo "- [CHECKED ✅] bandwhich command exists"
+fi
